Add -Ddse property for indicating DSE use.
If -Ddse=true is present, -Dcassandra.version will be interpreted as the
DSE version.

Added static method to CCMBridge for getCassandraVersion, which detects
the cassandra version from cassandra.version, or if DSE attempts to
derive the C* version.

@DseVersion annotation to restrict tests based on DSE version
requirement.  If DSE is not in use the test is simply skipped.